+++ client/src/components/landing/GenericSection.js
