+++ client/src/components/landing/Home.js
@@ -1,10 +1,10 @@
 import React from 'react';
 // import sections
-import Hero from '../components/sections/Hero';
-import FeaturesTiles from '../components/sections/FeaturesTiles';
-import FeaturesSplit from '../components/sections/FeaturesSplit';
-import Testimonial from '../components/sections/Testimonial';
-import Cta from '../components/sections/Cta';
+import Hero from '../sections/Hero';
+import FeaturesTiles from '../sections/FeaturesTiles';
+import FeaturesSplit from '../sections/FeaturesSplit';
+import Testimonial from '../sections/Testimonial';
+import Cta from '../sections/Cta';
 
 const Home = () => {
